Adds support for Terraform workspaces
The Docker container's name and external port have to be unique.  I can
include the workspace in the container's name to make it unique and let
Docker assign an unused port.

I was able to add an output for the container's name, but I couldn't
find a way to output the assigned port number.