Adds output of the port number
I found out how to output the assigned port number!

If I set the `server_port` variable to `0`, Docker will assign a port
number.  But `terraform apply` then sees that port number as different
than what's expected and a subsequent `terraform apply` will try to
recreate the entire resource as a consequence.  The only way I've found
to have it ignore the port is to leave the `external` bit out.